pyproject.toml
[project]
name = "resume-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "pywin32-ctypes>=0.2.3",
]
